Restructure Python module and add packaging metadata
