+++ python/ili2c/__init__.py
@@ -1,0 +1,8 @@
+"""Public Python API for the ili2c toolkit."""
+
+from .ilirepository import IliRepositoryManager, ModelMetadata  # re-export
+
+__all__ = [
+    "IliRepositoryManager",
+    "ModelMetadata",
+]
