+++ python/ili2c/pyili2c/metamodel/attribute.py
