+++ python/ili2c/pyili2c/metamodel/cardinality.py
