+++ python/ili2c/pyili2c/metamodel/domain.py
