+++ python/ili2c/pyili2c/metamodel/element.py
