+++ python/ili2c/pyili2c/metamodel/role.py
