+++ python/setup.cfg
@@ -1,0 +1,24 @@
+[metadata]
+name = ili2c-python
+version = 0.1.0
+description = Python utilities for ili2c (INTERLIS metamodel, parser, and repository helpers).
+long_description = file: README.md
+long_description_content_type = text/markdown
+author = ili2c Maintainers
+author_email = info@interlis.ch
+url = https://github.com/claeis/ili2c
+license = LGPL-2.1-or-later
+
+[options]
+packages = find:
+python_requires = >=3.10
+install_requires =
+    antlr4-python3-runtime>=4.13,<4.14
+
+[options.extras_require]
+test =
+    pytest>=8.0
+
+[options.packages.find]
+where = .
+include = ili2c*
